refactor: insert, update template 분리
